tests: verify that api surface is fully exposed
The SDK was sensitive to the fact that it was more or less required that the `__init__.py` files in `apimodels` modules/subpackages were manually updated. This change adds some reflection run during testing to catch cases where this was forgotten or overlooked.

Notably, there were in fact unexposed classes (from my own doing), justifying at least in part the benefit of this change.