fix: don't use context manager within simple client
This reflects a philosophical shift on my part. I previously would use a `with` statement with the specified horde client but I have come to the conclusion that it should be the caller's responsibility to do this at a time of their choosing. The status quo before this change would/could cause situations where requests would be cancelled in an almost certainly unintended manner. For example, if you had two concurrent simple requests going, the requests would race to finish and the first completing would cause the second to be cancelled.