fix: explicitly specify `default` arg for pydantic `Field`s
Either the behavior changed at some point or I overlooked the fact that the `default` arg is not propagated through to mypy properly if it is specified positionally. This makes it explicit in all cases to fix the bug, if there is one, and at the very least least clarifies the meaning of the argument.