refactor: `generation_ids` to `result_ids` rename
Due to the evolution of the design of the sdk, most of the usages of the `generation_id`(s) variables were confusingly named. In the current iteration, `generations` are instances of inference (or generating) , where `results` are the actual outputs from those. There can be 1...N results, but generations are always a single, discrete, entity from the perspective of the SDK.