tests: identify http verbs not in SDK or swagger
This closes a gap in testing where if an endpoint was supported in the SDK by any verb, but not all verbs, testing would still pass. This was specifically in response that a newly added DELETE verb to a previously existing endpoint (`/users/`) did not cause the testing to fail.